+++ apps/app/src/data/domain/lists/Trips.ts
